Fixing Artist tests :rocket:
